+++ test/fake/window_service.dart
@@ -11,6 +11,12 @@ class FakeWindowService extends Fake implements WindowService {
   Future<void> keepScreenOn(bool on) => SynchronousFuture(null);
 
   @override
+  Future<bool> isInMultiWindowMode() => SynchronousFuture(false);
+
+  @override
+  Future<bool> isInPictureInPictureMode() => SynchronousFuture(false);
+
+  @override
   Future<bool> isRotationLocked() => SynchronousFuture(false);
 
   @override
@@ -30,4 +36,7 @@ class FakeWindowService extends Fake implements WindowService {
 
   @override
   Future<void> setColorMode({required bool wideColorGamut, required bool hdr}) => SynchronousFuture(null);
+
+  @override
+  Future<bool> startGlobalDrag(String uri, String? label, Size shadowSize, Uint8List shadowBytes) => SynchronousFuture(false);
 }